# wfc/canvas/static/scripts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunPatchRequest": {
        "description": "Partial update for a Run's user-editable metadata.\n\nAll fields are optional; missing fields are left unchanged. `nid`\nwrites through to `Run.nid` (provenance table); the rest upsert a\n`RunAnnotation` row.\n\n`archived` is a bool at the API boundary and maps to the nullable\n`archived_at` timestamp column: ``true`` sets it to now, ``false``\nclears it.",
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "favorite": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorite"
          },
          "nid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nid"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "RunPatchRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Workflow Canvas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pipelines/{pipeline_id}/document": {
      "get": {
        "description": "Return the run's literal frozen ``pipeline.json``.\n\nReads ``<project_root>/.runs/pipelines/<pipeline_id>/pipeline.json`` \u2014\nwritten at submission time for canvas runs, frozen at run start by\n``run_pipeline`` for CLI runs, and the same file\n``WfcProvider._load_bundled_samples`` consumes for fan-in sample\nresolution. Returns the parsed JSON document as-is so the canvas can\nhand it to ``loadPipeline()`` without transformation.\n\n404 when the file does not exist (the pipeline was authored but\nnever reached the snake-gen / run-generation stage).",
        "operationId": "get_pipeline_document_api_pipelines__pipeline_id__document_get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Document"
      }
    },
...
    "/api/runs/{run_id}/lineage-pipeline": {
      "get": {
        "description": "Return a synthesized literal-only lineage pipeline for ``run_id``.\n\nWalks ``parentRunIds`` from the clicked run back to roots \u2014 through\npipeline boundaries \u2014 and synthesizes a flat literal-only pipeline\nJSON suitable for the canvas's ``loadPipeline()``. See\n``wfc.lineage.synthesis`` for the algorithm.\n\nStatus codes:\n  - 200 with synthesized JSON on success\n  - 404 if the run id is unknown\n  - 422 if synthesis fails (cycle defense, malformed ancestor chain)",
        "operationId": "get_run_lineage_pipeline_api_runs__run_id__lineage_pipeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Lineage Pipeline"
      }
    },
...
    "/api/wfc/run/{run_id}": {
      "get": {
        "operationId": "get_wfc_run_api_wfc_run__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Wfc Run"
      },
      "patch": {
        "description": "Partial-update a run's user-editable metadata.",
        "operationId": "patch_wfc_run_api_wfc_run__run_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Wfc Run"
      }
    },
...
    "/api/wfc/run/{run_id}/cancelled-descendants": {
      "get": {
        "description": "Return runs cancelled because ``run_id`` (or its subtree) failed.",
        "operationId": "get_wfc_cancelled_descendants_api_wfc_run__run_id__cancelled_descendants_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Wfc Cancelled Descendants"
      }
    },
...
    "/api/workflow/{pipeline_id}/editable": {
      "get": {
        "description": "Return the pre-substitution pipeline JSON (with variables + {$var}).\n\nFalls back to ``pipeline.json`` (post-substitution) for a run that has\nno editable sidecar.",
        "operationId": "get_pipeline_editable_api_workflow__pipeline_id__editable_get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Editable"
      }
    }
  }
}